+++ src/lib/actions/base.ts
@@ -203,7 +203,12 @@ export abstract class Action {
             const srcMod = mod as IWebpackStatsModuleSource;
             identifier = srcMod.identifier;
             name = srcMod.name;
-            size = srcMod.size;
+            // Note: there are isolated cases where webpack4 appears to be
+            // wrong in it's `size` estimation vs the actual string length.
+            // See `version mismatch for v1-v4 moment-app` wherein the
+            // real length of `moment/locale/es-us.js` is 3017 but webpack
+            // v4 reports it in stats object as 3029.
+            size = srcMod.source.length || srcMod.size;
             source = srcMod.source;
 
           } else if (RWebpackStatsModuleSynthetic.decode(mod).isRight()) {