+++ HISTORY.md
@@ -3,6 +3,7 @@ History
 
 ## UNRELEASED
 
+* Use `source` string length over `size` reported from Webpack stats for assing real size of source.
 * Upgrade prod and dev dependencies, including TypeScript (to `3.7.4`).
 
 ## 4.2.2