+++ HISTORY.md
@@ -1,6 +1,12 @@
 History
 =======
 
+## UNRELEASED
+
+* Use `source` string length over `size` reported from Webpack stats for assessing real size of source.
+* Remove `*.map` files from published npm package.
+* Upgrade prod and dev dependencies, including TypeScript (to `3.7.4`).
+
 ## 4.2.2
 
 * Update `yargs` for security.